+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Released on ??
 - Bugfix:
   - File mode of file on remote is now reported on local file after being downloaded (unix, linux, macos only)
   - Scp: when username was not provided, it didn't fallback to current username
+  - Explorer: fixed UID format in Windows
 
 ## 0.1.3